Sample transaction test using mocked history. Signs and verifies TX. (#28)
* Full key serialization impl. X509/PKCS8 and tests.

This is a relatively simple way of representing transaction data.
Mostly for building more tests. We can do the signing at any level
of abstraction. For simplicity just going to test signing on simple
json with the inputs (independent of any blockchain related data.)

Later on we can either sign blockchain data within the same data
structure, or just have a double-signing step. Can be modified easily
later.

* Refactoring, split transaction separate from wallet. Fix test names.

* Move function / optimize import. Minor refactor

* Sample transaction test using mocked history. Signs and verifies TX.

Long explanation of thought process going into a transaction included.
This demonstrates why double spending is an issue, and issues involving
PublicKey security. Creates a transaction and an inverse transaction
and performs a pseudo-lookup against a mocked network to verify.

* Add KeyPair CustomSerializers for json4s and tests. Wallet protoclass.